+++ Elements/src/Spatial/AdaptiveGrid/AdaptiveGrid.cs
@@ -320,10 +320,38 @@ namespace Elements.Spatial.AdaptiveGrid
             return TryGetValue(zDict, point.Z, out id, tolerance);
         }
 
+        /// <summary>
+        /// Add a Vertex and connect in to one or more other vertices.
+        /// </summary>
+        /// <param name="point">Position of required Vertex.</param>
+        /// <param name="connections">Ids of other vertices to connect new Vertex with.</param>
+        /// <returns>New Vertex or existing one if it's within grid tolerance.</returns>
+        public Vertex AddVertex(Vector3 point, List<Vertex> connections)
+        {
+            if (connections == null || !connections.Any())
+            {
+                throw new ArgumentException("Vertex should be connected to at least one other Vertex");
+            }
+
+            Vertex v = AddVertex(point);
+            foreach (var c in connections)
+            {
+                AddEdge(v.Id, c.Id);
+            }
+
+            return v;
+        }
+
         #endregion
 
         #region Private logic
 
+        /// <summary>
+        /// Add a Vertex or return existing one if it's withing grid tolerance.
+        /// Doesn't connect new Vertex to the grid with edges.
+        /// </summary>
+        /// <param name="point">Position of required vertex</param>
+        /// <returns>New or existing Vertex.</returns>
         private Vertex AddVertex(Vector3 point)
         {
             if (!TryGetVertexIndex(point, out var id, Tolerance))
@@ -339,6 +367,10 @@ namespace Elements.Spatial.AdaptiveGrid
             return GetVertex(id);
         }
 
+        /// <summary>
+        /// Remove the Vertex with specified id from the grid.
+        /// </summary>
+        /// <param name="id">Vertex id to delete.</param>
         private void DeleteVertex(ulong id)
         {
             var vertex = _vertices[id];
@@ -361,6 +393,12 @@ namespace Elements.Spatial.AdaptiveGrid
             }
         }
 
+        /// <summary>
+        /// Add an Edge or return the exiting one with given indexes.
+        /// </summary>
+        /// <param name="vertexId1">Index of the first Vertex</param>
+        /// <param name="vertexId2">Index of the second Vertex</param>
+        /// <returns>New or existing Edge.</returns>
         private Edge AddEdge(ulong vertexId1, ulong vertexId2)
         {
             if (vertexId1 == vertexId2)
@@ -391,6 +429,10 @@ namespace Elements.Spatial.AdaptiveGrid
             }
         }
 
+        /// <summary>
+        /// Remove the Edge from the grid.
+        /// </summary>
+        /// <param name="edge">Edge to delete</param>
         private void DeleteEdge(Edge edge)
         {
             var hash = Edge.GetHash(new List<ulong> { edge.StartId, edge.EndId });