+++ Elements/test/AdaptiveGridTests.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using Xunit;
+using Vertex = Elements.Spatial.AdaptiveGrid.Vertex;
 
 namespace Elements.Tests
 {
@@ -35,7 +36,6 @@ namespace Elements.Tests
                 Model.AddElement(new ModelCurve(edge.GetGeometry(), material: random.NextMaterial()));
             }
         }
-
 
         [Fact]
         public void AdaptiveGridBboxKeyPointsExample()
@@ -76,6 +76,34 @@ namespace Elements.Tests
             {
                 Model.AddElement(new ModelCurve(edge.GetGeometry(), material: random.NextMaterial()));
             }
+        }
+
+        [Fact]
+        public void AdaptiveGridAddVertex()
+        {
+            var adaptiveGrid = new AdaptiveGrid(new Transform());
+            var points = new List<Vector3>()
+            {
+                new Vector3(-6, -4),
+                new Vector3(-2, -4),
+                new Vector3(3, -4),
+                new Vector3(1, 4.5, 3),
+                new Vector3(6, 3, -2),
+            };
+            adaptiveGrid.AddFromPolygon(Polygon.Rectangle(15, 10), points);
+
+            ulong id;
+            Assert.True(adaptiveGrid.TryGetVertexIndex(new Vector3(-2, -4), out id));
+            var oldV = adaptiveGrid.GetVertex(id);
+            var edgesBefore = oldV.Edges.Count;
+
+            var newV = adaptiveGrid.AddVertex(new Vector3(-2, -4, 2), new List<Vertex> { oldV });
+            Assert.NotNull(newV);
+            Assert.False(newV.Id == 0);
+            Assert.Single(newV.Edges);
+            Assert.True(newV.Edges.First().StartId == id || newV.Edges.First().EndId == id);
+            Assert.Equal(edgesBefore + 1, oldV.Edges.Count());
+            Assert.Contains(oldV.Edges, e => e.StartId == newV.Id || e.EndId == newV.Id);
         }
 
         [Fact]