+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - `ModelLines`
+- `AdaptiveGrid.AddVertex(Vector3 point, List<Vertex> connections)`
 
 ### Changed
 - Add parameter `removeCutEdges` to `AdaptiveGrid.SubtructBox` that control if cut parts of intersected edges need to be inserted into the graph.